Module rustpython_parser::ast

Implement abstract syntax tree (AST) nodes for the python language.

Roughly equivalent to the python AST Many AST nodes have a location attribute, to determine the sourcecode location of the node.

Re-exports

pub use crate::location::Location;

Structs

ArgumentList

Comprehension
A list/set/dict/generator compression.

ExceptHandler

ImportSymbol

Keyword

Located

Parameter
A single formal parameter to a function.

Parameters
Formal parameters to a function.

Program
A full python program, it's a sequence of statements.

WithItem

Enums

BooleanOperator
A boolean operation.

Comparison
A comparison operation.

ComprehensionKind

ConversionFlag
Transforms a value prior to formatting it.

ExpressionType
A certain type of expression.

Number
A numeric literal.

Operator
An operator for a binary operation (an operation with two operands).

StatementType
Abstract syntax tree nodes for python statements.

StringGroup

Top

UnaryOperator
An unary operator. This is an operation with only a single operand.

Varargs

Type Definitions

Expression
An expression at a given location in the sourcecode.

Statement

Suite