Type Definition rustpython_parser::ast::Expression

type Expression = Located<ExpressionType>;

An expression at a given location in the sourcecode.

Methods

impl Expression

pub fn name(&self) -> &'static str

Returns a short name for the node suitable for use in error messages.